==> 06A_transfer_learning_model.ipynb <==
Basic version already sort of done in Collab

Transfer model with VGG and ResNEt
Choose 1 or can keep both types

!pip install --upgrade keras-cv tensorflow
!pip install --upgrade keras

import tensorflow as tf
from tensorflow import keras
import keras_cv
from keras_cv import layers as cv_layers
import numpy as np
import pandas as pd
from tqdm import tqdm
import pickle
import glob
import cv2
import re
import os
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split

# Check version
print(tf.__version__)
print(keras.__version__)
print(keras_cv.__version__)

# Connect to Drive folder with Data
from google.colab import drive
drive.mount('/content/drive')
folder_path = '/content/drive/My Drive/Sarah_Master_Project/Data/Colab_data'

#TODO: this is for Colab obviously change this if in vscode

image_file = os.path.join(folder_path, 'Kenya2019_array_augmented.npy')
images = np.load(image_file, allow_pickle=True)
images = images.item()


df = pd.read_csv(os.path.join(folder_path, 'clustered_geopoints.txt'), sep = '\t')
#labels = df['avg_wealthscore'].values
df['avg_roofcat'] = df['avg_roofcat'].replace(0.5, 0)
labels = df['avg_roofcat'].values
image_codes = df['image_code'].astype(str).values

#Remove nan
non_nan = ~np.isnan(labels)
labels = labels[non_nan]
image_codes = image_codes[non_nan]

df['avg_roofcat'].value_counts()

# Normalise images
images = {key: np.array(img).astype('float32') / 255.0 for key, img in images.items()}

X = np.array([images[code] for code in image_codes if code in images])


print(f"Image data shape: {X.shape}")
print(f"Labels shape: {labels.shape}")

X_train, X_test, y_train, y_test = train_test_split(X, labels, test_size=0.2, random_state=42)


# Flat arrays
X_train_flat = X_train.reshape(X_train.shape[0], -1)
X_test_flat = X_test.reshape(X_test.shape[0], -1)


# Preprocess
NUM_CLASSES = 2
BATCH_SIZE = 16


def preprocess_data(images, labels, augment=False):
    labels = tf.cast(labels, tf.int32)
    if NUM_CLASSES > 1:  
        labels = tf.one_hot(labels, NUM_CLASSES)
    inputs = {"images": images, "labels": labels}
    outputs = inputs
    return outputs['images'], outputs['labels']

# Convert your images and labels to TensorFlow datasets
train_dataset = tf.data.Dataset.from_tensor_slices((X_train, y_train))
test_dataset = tf.data.Dataset.from_tensor_slices((X_test, y_test))

# Apply preprocessing, batching, and prefetching to datasets
train_dataset = train_dataset.batch(BATCH_SIZE).map(
    lambda x, y: preprocess_data(x, y, augment=True),
    num_parallel_calls=tf.data.AUTOTUNE
).prefetch(tf.data.AUTOTUNE)

test_dataset = test_dataset.batch(BATCH_SIZE).map(
    lambda x, y: preprocess_data(x, y, augment=False),
    num_parallel_calls=tf.data.AUTOTUNE
).prefetch(tf.data.AUTOTUNE)


# Model: RESNET 50

For vgg change backbone

backbone = keras_cv.models.EfficientNetV2Backbone.from_preset("efficientnetv2_b0_imagenet")


backbone = keras_cv.models.ResNetBackbone.from_preset("resnet50_imagenet")


if NUM_CLASSES > 1:  # For classification
    activation = "softmax"
    loss = 'categorical_crossentropy'
else:  # For regression
    activation = "linear"
    loss = 'mean_squared_error'

model = keras_cv.models.ImageClassifier(
    backbone=backbone,
    num_classes=NUM_CLASSES,
    activation=activation,
)

model.compile(
    loss=loss,
    optimizer=keras.optimizers.Adam(learning_rate=1e-5),
    metrics=['accuracy'] if NUM_CLASSES > 1 else ['mae']
)

model.fit(
    train_dataset,
    validation_data=test_dataset,
    epochs=16,
    verbose =1
)

model.evaluate(test_dataset)

from sklearn.metrics import mean_absolute_error

y_pred = model.predict(test_dataset)
y_pred = np.argmax(y_pred, axis=1)
mae = mean_absolute_error(y_test, y_pred)
print(f'Mean Absolute Error: {mae}')

print(y_pred.shape)
print(y_test.shape)

from sklearn.metrics import r2_score

r2 = r2_score(y_test, y_pred)
print(f'R-squared: {r2}')


from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred))
